refactor: fix eslint
