+++ src/components/Flow/index.js
@@ -19,83 +19,83 @@ class Flow extends React.Component {
     this.containerId = `${FLOW_CONTAINER_ID}_${uuid()}`;
   }
 
-    canDragCanvas = () => {
-      const { labelState } = this.props;
+  canDragCanvas = () => {
+    const { labelState } = this.props;
 
-      return labelState === LABEL_STATE_HIDE;
-    };
+    return labelState === LABEL_STATE_HIDE;
+  };
 
-    canZoomCanvas = () => {
-      const { labelState } = this.props;
+  canZoomCanvas = () => {
+    const { labelState } = this.props;
 
-      return labelState === LABEL_STATE_HIDE;
-    };
+    return labelState === LABEL_STATE_HIDE;
+  };
 
-    canDragNode = ({ target }) => {
-      return target && target.get('className') !== SHPAE_CLASSNAME_ANCHOR;
-    };
+  canDragNode = ({ target }) => {
+    return target && target.get('className') !== SHPAE_CLASSNAME_ANCHOR;
+  };
 
-    parseData = ({ data }) => {
-      const { nodes, edges } = data;
+  parseData = ({ data }) => {
+    const { nodes, edges } = data;
 
-      [...nodes, ...edges].forEach((item) => {
-        const { id } = item;
+    [...nodes, ...edges].forEach((item) => {
+      const { id } = item;
 
-        if (id) {
-          return;
-        }
+      if (id) {
+        return;
+      }
 
-        item.id = uuid();
-      });
-    };
+      item.id = uuid();
+    });
+  };
 
-    initGraph = ({ width, height }) => {
-      const { containerId } = this;
+  initGraph = ({ width, height }) => {
+    const { containerId } = this;
 
-      this.graph = new G6.Graph({
-        container: containerId,
-        width,
-        height,
-        modes: {
-          default: [
-            {
-              type: 'drag-canvas',
-              shouldBegin: this.canDragCanvas,
-              shouldUpdate: this.canDragCanvas,
-              shouldEnd: this.canDragCanvas,
-            },
-            {
-              type: 'zoom-canvas',
-              shouldUpdate: this.canZoomCanvas,
-            },
-            'click-node',
-            'hover-node',
-            'hover-anchor',
-            {
-              type: 'drag-node',
-              shouldBegin: this.canDragNode,
-            },
-            'drag-add-edge',
-            'edit-label',
-          ],
-        },
-      });
+    this.graph = new G6.Graph({
+      container: containerId,
+      width,
+      height,
+      modes: {
+        default: [
+          {
+            type: 'drag-canvas',
+            shouldBegin: this.canDragCanvas,
+            shouldUpdate: this.canDragCanvas,
+            shouldEnd: this.canDragCanvas,
+          },
+          {
+            type: 'zoom-canvas',
+            shouldUpdate: this.canZoomCanvas,
+          },
+          'click-node',
+          'hover-node',
+          'hover-anchor',
+          {
+            type: 'drag-node',
+            shouldBegin: this.canDragNode,
+          },
+          'drag-add-edge',
+          'edit-label',
+        ],
+      },
+    });
 
-      return this.graph;
-    };
+    return this.graph;
+  };
 
-    render() {
-      const { containerId, parseData, initGraph } = this;
+  render() {
+    const { containerId, parseData, initGraph } = this;
 
-      return (
-        <Graph
-          containerId={containerId}
-          parseData={parseData}
-          initGraph={initGraph}
-          {...this.props}
-        />
-      );
-    }
+    return (
+      <Graph
+        containerId={containerId}
+        parseData={parseData}
+        initGraph={initGraph}
+        {...this.props}
+      />
+    );
+  }
 }
 
 export default withEditorContext(Flow, ({ labelState }) => ({
